Address nit feedback
