+++ copilot/fsd-pre-award-send-application-assessment-report/manifest.yml
@@ -9,7 +9,7 @@ timeout: 10m
 # We inject `image.location` in our deployment pipeline
 image: {}
 entrypoint: launcher
-command: invoke reporting.send-report
+command: invoke reporting.send-application-assessment-report
 
 cpu: 256
 memory: 512
